AMC: Add cvar `cl_keybindmode` to change keyboard config behavior
If 0, will allow multiple gamefuncs to be assigned to the same key
in the keyboard config menu (old behavior).

If 1, will clear all existing gamefuncs for that key when binding a
key to a gamefunc. This prevents accidentally assigning multiple
gamefuncs to the same key, e.g. when default values exist.