AMC: Implement a DEF- and CON-controllable custom settings menu.
* New DEF token: `customsettings { ... }`
Can be used to define the structure of the custom settings menu.
Maximum of 64 entries.

* New events:
"EVENT_CSACTIVATELINK", "EVENT_CSPREMODIFYOPTION",
"EVENT_CSPOSTMODIFYOPTION", "EVENT_CSPOPULATEMENU"
Executed on activating a link, modifying an option, and when
the menu is to be set up (populated).

* New userdef structs:
ud.csarray - array that stores the settings altered by the menu.
ud.m_customsettings - menu index for the setting

Also added a lower bound for distance between menu items.
Documentation on their usage will follow in a later commit