AMC: Add PROJECTILE_WORKSLIKE flag 'PROJECTILE_RADIUS_PICNUM_EX'
Flag value is 1073741824 (0x40000000).

The flag changes the htpicnum of the projectile to the actual picnum,
instead of using RADIUSEXPlOSION. However, it also preserves hardcoded
behaviors specific to RADIUSEXPLOSION, including destroying spritewalls
and spritefloors that have a hitag.

This commit also makes it so that RPG projectiles can only burn trees,
tires and boxes if this flag is set. Moreover, with this flag set, the
player will receive the same pushback as with RADIUSEXPLOSION when hit.