Engine: fix regressions with clipupdatesector() from df868267
This resolves a bug in Duke3D where the player couldn't move
while shrunk if a TROR sector was present on the map.

It also ensures that the original clipping is used if no TROR
bunch is present inside the sector in question. This should
reduce the amount of backwards-compatibility problems.